update folder structure & finish user login
