+++ src/components/content/add_data/index.js
@@ -1,33 +1,32 @@
 import React from 'react';
-import { useDispatch } from 'react-redux';
-import { ACTIVITY_TYPE, ACTIVITY_ICONS } from '../../../helpers/constant';
-import { setCurrentPage } from '../../../slices/statusSlice';
+import { useSelector } from 'react-redux';
+import AddDataHome from './addDataHome';
+import AddActivity from './addActivity';
 
-const AddDataHome = () => {
-  const dispatch = useDispatch();
+const AddDataPage = () => {
+  const { currentPage } = useSelector(state => state.status);
+
+  const selectPage = current => {
+    const activity = current.split(' ')[1];
+    switch (current) {
+      case 'add walking':
+      case 'add running':
+      case 'add cycling':
+      case 'add swimming':
+      case 'add exercise':
+        return <AddActivity activity={activity} />;
+      default:
+        return <AddDataHome />;
+    }
+  };
+
+  const page = selectPage(currentPage);
 
   return (
-    <div className="addData-home">
-      {
-        ACTIVITY_TYPE.map(activity => (
-          <button
-            type="button"
-            key={activity}
-            className="addData-card"
-            onClick={() => dispatch(setCurrentPage(`add ${activity}`))}
-          >
-            <div className="icon">
-              {ACTIVITY_ICONS(activity)}
-            </div>
-            <div className="title">
-              {activity}
-            </div>
-          </button>
-        ))
-      }
+    <div className="addData-page">
+      {page}
     </div>
   );
 };
 
-
-export default AddDataHome;
+export default AddDataPage;
